requirements/base.txt
psycopg2-binary==2.9.10
typing_extensions==4.12.2
whitenoise==6.9.0  # https://github.com/evansd/whitenoise
python-decouple==3.8
langchain==0.3.19
langchain-openai==0.3.6
langchain-core==0.3.35
langchain-community==0.3.17
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
Pillow==11.1.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
redis==5.2.1  # https://github.com/redis/redis-py
hiredis==3.1.0  # https://github.com/redis/hiredis-py
celery==5.4.0  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.7.0  # https://github.com/celery/django-celery-beat

# Django
# -----------------------------------------------------------------

Django==5.1.6
django-allauth==65.4.1  # https://github.com/pennersr/django-allauth
django-environ==0.12.0
django-redis==5.4.0

# Django REST Framework
# -----------------------------------------------------------------

django-cors-headers==4.7.0
djangorestframework==3.15.2
drf-yasg==1.21.8
dj-rest-auth==7.0.1
django-extensions==3.2.3
djangorestframework-simplejwt==5.4.0
django-anymail==12.0
pytest-django==4.10.0
celery==5.4.0
